Cleaning ShopItem objects. Adjusted paths from absolute to relative.
