+++ run_tests.py
@@ -1,13 +1,6 @@
 import datetime
 
-from src.browser.scraper_bots.argos.argos import Argos
-from src.browser.scraper_bots.boots.boots import Boots
-from src.browser.scraper_bots.debenhams.debenhams_watches import DebenhamsWatchesShop
-from src.browser.scraper_bots.fragranceshop.fragranceshop import FragranceShop
-from src.browser.scraper_bots.hsamuel.hsamuel import HSamuelShop
-from src.browser.scraper_bots.johnlewis.johnlewis_watches import JohnLewisWatchesShop
-from src.browser.scraper_bots.watches2u.watches2u import Watches2uShop
-from src.browser.scraper_bots.watchshop.watchshopshop import WatchShopShop
+from src.browser.scraper_bots.argos import Argos
 from src.browser.webdriver import WebDriver
 
 web_driver = WebDriver()
