+++ src/browser/scraper_bots/debenhams_watches.py
@@ -1,14 +1,11 @@
-import datetime
 from time import sleep
 
-from selenium.common.exceptions import NoSuchElementException
-from selenium.webdriver import ActionChains
 from selenium.webdriver.common.keys import Keys
 
-from src.browser.scraper_bots.debenhams.debenhams_item import DebenhamsItem
+from src.browser.scraper_bots.ShopItem import ShopItem
 
 
-class JohnLewisWatchesShop(object):
+class DebenhamsWatchesShop(object):
     def __init__(self, driver, store_id, date, time):
         self.driver = driver
         self.store_id = store_id
@@ -17,27 +14,19 @@ class JohnLewisWatchesShop(object):
 
     @property
     def item_boxes(self):
-        class_name = 'product-card'
-        self.driver.find_element_by_tag_name('html').send_keys(Keys.END)
-        sleep(20)
+        class_name = 't-product-list__product'
         elements = self.driver.find_elements_by_class_name(class_name)
         print(len(elements))
-        return elements
-        # for item in elements:
-        #     yield item
+        for item in elements:
+            yield item
 
     def scrape_products(self):
 
         # CSS paths
         name_css = 'div > article > div > div.u-display-inline-block.u-width-full > a > div'
         prices_css = 'div > article > div > div:nth-child(2) > div.dbh-now.u-text-weight-bold.u-color-neutral-70 > span'
-        price_class = ''
-        prices_css_2 = ''
-        url_class = ''
-
         image_css = 'div > article > a > div > img'
         href_css = 'div > article > a'
-        promo_class = ''
 
         for item in self.item_boxes:
             try:
@@ -49,8 +38,8 @@ class JohnLewisWatchesShop(object):
                 promo = None
                 promo_url = None
 
-                item = DebenhamsItem(title, price, store_product_id, url, item_image, self.store_id, self.date, self.time, promo, promo_url)
-                print(item)
+                item = ShopItem(title, price, store_product_id, url, item_image, self.store_id, self.date, self.time,
+                                promo, promo_url)
             except:
                 continue
             yield item
@@ -73,16 +62,7 @@ class JohnLewisWatchesShop(object):
         while True:
             for item in self.scrape_products():
                 continue
-                #item.evaluate()
-            #break
             self.nex_page()
-            # if self.next_page() is not None:
-            #     print(self.next_page())
-            #     self.driver.get(self.next_page())
-            #     now = datetime.datetime.now()
-            #     self.date, self.time = now.strftime("%Y-%m-%d"), now.strftime("%H:%M")
-            # elif self.next_page() is None:
-            #     break
 
     def test_run2(self):
         while True:
@@ -100,18 +80,10 @@ class JohnLewisWatchesShop(object):
 
     def nex_page(self):
         button_next_page = '#app-main > div > div > div.t-product-list__side-filter-container.t-product-list__side-filter-container-no-results-container > div:nth-child(2) > div > div.t-product-list__result-list-wrapper > div.t-product-list__container.dbh-product-list > div.c-product-control-bar.c-product-control-bar--bottom > div > nav > button.pw-button.pw--icon-only.pw-pagination__next.pw-pagination__button.dbh-next.u-flexbox.u-align-center'
-
-        # self.driver.find_element_by_tag_name('html').send_keys(Keys.END)
-        # sleep(0.5)
         pagination = self.driver.find_element_by_css_selector(button_next_page)
 
         if pagination.is_enabled():
             return pagination
         else:
             return None
-            # element = next_page.find_element_by_tag_name('a').get_attribute('href')
-            # next_page.find_element_by_tag_name('a').click()
 
-        # except NoSuchElementException:
-        #     element = None
-        #     return element
