--- src/browser/scraper_bots/ernestjones/__init__.py
@@ -1,3 +1,0 @@
-
-__author__ = "martin"
-
