+++ src/browser/scraper_bots/hsamuel.py
@@ -1,17 +1,12 @@
 import datetime
 from time import sleep
 
-from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver import ActionChains
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.keys import Keys
 
-from src.browser.scraper_bots.boots.boots_item import BootsItem
-from src.browser.scraper_bots.ernestjones.ernestjones_item import ErnestJonesItem
-from src.browser.scraper_bots.hsamuel.hsamuel_item import HSamuelItem
+from src.browser.scraper_bots.ShopItem import ShopItem
 
 
-class ErnestJonesShop(object):
+class HSamuelShop(object):
     def __init__(self, driver, store_id, date, time):
         self.driver = driver
         self.store_id = store_id
@@ -23,42 +18,42 @@ class ErnestJonesShop(object):
         div_class = 'items'
         '#js-product-list'
         element = self.driver.find_element_by_class_name(div_class)
-        #print(len(element.find_elements_by_tag_name("li")))
         for item in element.find_elements_by_tag_name("li"):
             yield item
 
     def scrape_products(self):
 
         # CSS paths
-        stock_css = 'div > div > a > div:nth-child(2) > div.tablet-up > div > span'
+        stock_css = 'div > div > a > div:nth-child(2) > div.tablet-up > span'
         image_css = 'div > a.productLink.link--plain.js-nextPreviousStorage > div > img'
         title_css = 'div > div > a > div.product-tile__description'
         sku_css = 'div > meta'
         price_css = 'div > div > a > div:nth-child(2) > div.product-tile__pricing-container > div > div > p > span > span.product-tile__price-value'
         url_css = 'div > div > a'
         for item in self.item_boxes:
-            stock = 'In Stock'
             try:
-                try:
-                    stock = item.find_element_by_css_selector(stock_css).text
-                except NoSuchElementException:
-                    if stock != 'Out of stock':
-                        title = item.find_element_by_css_selector(title_css).text.strip()
+                stock = item.find_element_by_css_selector(stock_css).text
+                if stock == 'In stock':
+                    title = item.find_element_by_css_selector(title_css).text.strip()
 
-                        if item.find_element_by_css_selector(image_css).get_attribute('src')[:5] == 'https':
-                            item_image = item.find_element_by_css_selector(image_css).get_attribute('src')
-                        else:
-                            item_image = None
+                    if item.find_element_by_css_selector(image_css).get_attribute('src')[:5] == 'https':
+                        item_image = item.find_element_by_css_selector(image_css).get_attribute('src')
+                    else:
+                        item_image = None
 
-                        store_product_id = item.find_element_by_css_selector(sku_css).get_attribute('content')
-                        price = float(item.find_element_by_css_selector(price_css).text)
-                        url = item.find_element_by_css_selector(url_css).get_attribute('href')
-                        promo, promo_url = None, None
-                        item = ErnestJonesItem(title, price, store_product_id, url, item_image, self.store_id, self.date, self.time, promo, promo_url)
-                        #print(item)
-                        yield item
+                    store_product_id = item.find_element_by_css_selector(sku_css).get_attribute('content')
+                    price = float(item.find_element_by_css_selector(price_css).text)
+                    url = item.find_element_by_css_selector(url_css).get_attribute('href')
+                    promo, promo_url = None, None
+                    item = ShopItem(title, price, store_product_id, url, item_image, self.store_id, self.date,
+                                    self.time, promo, promo_url)
+                    #print(item)
+                    yield item
             except:
                 continue
+            # item = HSamuelItem(title, price, store_product_id, url, item_image, self.store_id, self.date, self.time, promo, promo_url)
+            # print(item)
+            # yield item
 
     def process_items(self):
         close_button_css = '#js-cancel'
@@ -88,31 +83,9 @@ class ErnestJonesShop(object):
             elif self.next_page() is None:
                 break
 
-    def test2(self):
-        if self.cur_page_number() == 1:
-            while self.cur_page_number() < 4:
-                button = self.driver.find_element_by_css_selector('#js-load-next')
-                actions = ActionChains(self.driver)
-                actions.move_to_element(button)
-                actions.click(button)
-                actions.perform()
-                sleep(0.5)
-
-        while True:
-            for item in self.scrape_products():
-                continue
-                # item.evaluate()
-            if self.next_page() is not None:
-                print(self.next_page())
-                self.driver.get(self.next_page())
-                now = datetime.datetime.now()
-                self.date, self.time = now.strftime("%Y-%m-%d"), now.strftime("%H:%M")
-            elif self.next_page() is None:
-                break
-
     def test_run(self):
         if self.cur_page_number() == 1:
-            while self.cur_page_number() < 4:
+            while self.cur_page_number() < 3:
                 button = self.driver.find_element_by_css_selector('#js-load-next')
                 actions = ActionChains(self.driver)
                 actions.move_to_element(button)
@@ -150,18 +123,6 @@ class ErnestJonesShop(object):
         else:
             return None
 
-        # button_css = '#js-load-next'
-        #
-        # button = self.driver.find_element_by_css_selector(button_css)
-        # if button.get_attribute('class') == 'load-next':
-        #     current_url = self.driver.current_url
-        #     current_page = int(current_url.split('=')[-1])
-        #     next_page = current_page + 1
-        #     next_page = self.split(current_url, '=', 2) + '=' + str(next_page)
-        #     return next_page
-        # else:
-        #     return None
-
     def nex_page_once(self):
         button_css = '#js-load-next'
         close_button_css = '#js-cancel'
@@ -179,8 +140,3 @@ class ErnestJonesShop(object):
                 actions.click(button)
                 actions.perform()
                 sleep(0.5)
-
-    @staticmethod
-    def split(string, sep, pos):
-        string = string.split(sep)
-        return sep.join(string[:pos])# , sep.join(string[pos:])
