--- src/browser/scraper_bots/watchshop/__init__.py
@@ -1,3 +1,0 @@
-
-__author__ = "martin"
-
